web_admin/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
nltk==3.6.2
psycopg2-binary==2.9.5
requests==2.26.0
python-multipart==0.0.5
pytest==7.2.2